+++ cli/actions/run.go
@@ -1,0 +1,39 @@
+/*
+Copyright 2022 Nethermind
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package actions
+
+import (
+	"path/filepath"
+
+	"github.com/NethermindEth/sedge/configs"
+	"github.com/NethermindEth/sedge/internal/pkg/commands"
+	log "github.com/sirupsen/logrus"
+)
+
+type RunContainersOptions struct {
+	GenerationPath string
+	Services       []string
+}
+
+func (s *sedgeActions) RunContainers(options RunContainersOptions) error {
+	upCmd := s.commandRunner.BuildDockerComposeUpCMD(commands.DockerComposeUpOptions{
+		Path:     filepath.Join(options.GenerationPath, configs.DefaultDockerComposeScriptName),
+		Services: options.Services,
+	})
+	log.Infof(configs.RunningCommand, upCmd.Cmd)
+	_, err := s.commandRunner.RunCMD(upCmd)
+	return err
+}