+++ internal/pkg/generate/generate_scripts.go
@@ -19,6 +19,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
 	"text/template"
 
 	"github.com/NethermindEth/sedge/configs"
@@ -118,7 +119,7 @@ a. error
 Error if any
 */
 func generateDockerComposeScripts(gd GenerationData) (dockerComposePath string, err error) {
-	rawBaseTmp, err := templates.Services.ReadFile(filepath.Join("services", "docker-compose_base.tmpl"))
+	rawBaseTmp, err := templates.Services.ReadFile(strings.Join([]string{"services", "docker-compose_base.tmpl"}, "/"))
 	if err != nil {
 		return
 	}
@@ -135,14 +136,15 @@ func generateDockerComposeScripts(gd GenerationData) (dockerComposePath string, 
 	}
 	for tmpKind, client := range clients {
 		name := client.Name
-		if client.Omited {
+		if client.Omitted {
 			name = "empty"
 		}
-		tmp, err := templates.Services.ReadFile(filepath.Join("services",
-			configs.NetworksConfigs[gd.Network].NetworkService,
+		tmp, err := templates.Services.ReadFile(strings.Join([]string{
+			"services",
+			configs.NetworkConfigs()[gd.Network].NetworkService,
 			tmpKind,
-			name+".tmpl",
-		))
+			name + ".tmpl",
+		}, "/"))
 		if err != nil {
 			return "", err
 		}
@@ -150,6 +152,15 @@ func generateDockerComposeScripts(gd GenerationData) (dockerComposePath string, 
 		if err != nil {
 			return "", err
 		}
+	}
+
+	// Parse validator-blocker template
+	tmp, err := templates.Services.ReadFile(filepath.Join("services", "validator-blocker.tmpl"))
+	if err != nil {
+		return "", err
+	}
+	if _, err := baseTmp.Parse(string(tmp)); err != nil {
+		return "", err
 	}
 
 	// Check for TTD in env base template
@@ -215,6 +226,7 @@ func generateDockerComposeScripts(gd GenerationData) (dockerComposePath string, 
 	}
 
 	data := DockerComposeData{
+		Services:            gd.Services,
 		TTD:                 TTD,
 		CcCustomCfg:         ccRemoteCfg || ccRemoteGen || ccRemoteDpl,
 		CcRemoteCfg:         ccRemoteCfg,
@@ -249,6 +261,7 @@ func generateDockerComposeScripts(gd GenerationData) (dockerComposePath string, 
 		SplittedNetwork:     splittedNetwork,
 		ClCheckpointSyncUrl: clCheckpointSyncUrl,
 		LoggingDriver:       gd.LoggingDriver,
+		VLStartGracePeriod:  gd.VLStartGracePeriod,
 	}
 
 	dockerComposePath = filepath.Join(gd.GenerationPath, configs.DefaultDockerComposeScriptName)
@@ -282,7 +295,7 @@ a. error
 Error if any
 */
 func generateEnvFile(gd GenerationData) (envFilePath string, err error) {
-	rawBaseTmp, err := templates.Envs.ReadFile(filepath.Join("envs", gd.Network, "env_base.tmpl"))
+	rawBaseTmp, err := templates.Envs.ReadFile(strings.Join([]string{"envs", gd.Network, "env_base.tmpl"}, "/"))
 	if err != nil {
 		return
 	}
@@ -299,18 +312,18 @@ func generateEnvFile(gd GenerationData) (envFilePath string, err error) {
 	}
 	for tmpKind, client := range clients {
 		var tmp []byte
-		if client.Omited {
-			tmp, err = templates.Services.ReadFile(filepath.Join(
+		if client.Omitted {
+			tmp, err = templates.Services.ReadFile(strings.Join([]string{
 				"services",
-				configs.NetworksConfigs[gd.Network].NetworkService,
+				configs.NetworkConfigs()[gd.Network].NetworkService,
 				tmpKind,
 				"empty.tmpl",
-			))
+			}, "/"))
 			if err != nil {
 				return "", err
 			}
 		} else {
-			tmp, err = templates.Envs.ReadFile(filepath.Join("envs", gd.Network, tmpKind, client.Name+".tmpl"))
+			tmp, err = templates.Envs.ReadFile(strings.Join([]string{"envs", gd.Network, tmpKind, client.Name + ".tmpl"}, "/"))
 			if err != nil {
 				return "", err
 			}
@@ -324,11 +337,11 @@ func generateEnvFile(gd GenerationData) (envFilePath string, err error) {
 	// TODO: Use OS wise delimiter for these data structs
 	data := EnvData{
 		ElImage:                   gd.ExecutionClient.Image,
-		ElDataDir:                 configs.ExecutionDefaultDataDir,
+		ElDataDir:                 "./" + configs.ExecutionDir,
 		CcImage:                   gd.ConsensusClient.Image,
-		CcDataDir:                 configs.ConsensusDefaultDataDir,
+		CcDataDir:                 "./" + configs.ConsensusDir,
 		VlImage:                   gd.ValidatorClient.Image,
-		VlDataDir:                 configs.ValidatorDefaultDataDir,
+		VlDataDir:                 "./" + configs.ValidatorDir,
 		ExecutionApiURL:           gd.ExecutionClient.Endpoint + ":" + gd.Ports["ELApi"],
 		ExecutionAuthURL:          gd.ExecutionClient.Endpoint + ":" + gd.Ports["ELAuth"],
 		ConsensusApiURL:           gd.ConsensusClient.Endpoint + ":" + gd.Ports["CLApi"],
@@ -337,7 +350,7 @@ func generateEnvFile(gd GenerationData) (envFilePath string, err error) {
 		JWTSecretPath:             gd.JWTSecretPath,
 		ExecutionEngineName:       gd.ExecutionClient.Name,
 		ConsensusClientName:       gd.ConsensusClient.Name,
-		KeystoreDir:               configs.KeystoreDefaultDataDir,
+		KeystoreDir:               "./" + configs.KeystoreDir,
 		Graffiti:                  gd.Graffiti,
 	}