+++ internal/pkg/services/image_test.go
@@ -1,0 +1,56 @@
+package services_test
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/NethermindEth/sedge/internal/pkg/services"
+	mock_client "github.com/NethermindEth/sedge/test/mock_docker"
+	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/errdefs"
+	"github.com/golang/mock/gomock"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestImageFound(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	dockerClient := mock_client.NewMockAPIClient(ctrl)
+	defer ctrl.Finish()
+
+	expectedImage := "expected-image"
+
+	dockerClient.EXPECT().
+		ContainerInspect(context.Background(), "validator-client").
+		Return(types.ContainerJSON{
+			ContainerJSONBase: &types.ContainerJSONBase{
+				Image: expectedImage,
+			},
+		}, nil).
+		Times(1)
+
+	serviceManager := services.NewServiceManager(dockerClient)
+	image, err := serviceManager.Image("validator-client")
+	assert.Nil(t, err)
+	assert.Equal(t, expectedImage, image)
+}
+
+func TestImageNotFound(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	dockerClient := mock_client.NewMockAPIClient(ctrl)
+	defer ctrl.Finish()
+
+	expectedError := errdefs.NotFound(fmt.Errorf("error"))
+
+	dockerClient.EXPECT().
+		ContainerInspect(context.Background(), "validator-client").
+		Return(types.ContainerJSON{}, expectedError).
+		Times(1)
+
+	serviceManager := services.NewServiceManager(dockerClient)
+	image, err := serviceManager.Image("validator-client")
+	assert.ErrorIs(t, err, expectedError)
+	assert.True(t, errdefs.IsNotFound(err))
+	assert.Equal(t, "", image)
+}