+++ CHANGELOG.md
@@ -10,12 +10,42 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Documentation for all supported networks and clients on each of them.
-- Support for Erigon as Execution Client
-- Support for Besu as Execution Client
+- When generating new mnemonic, show it without a trace.
+- Support import/export slashing protection interchange data [EIP-3076](https://eips.ethereum.org/EIPS/eip-3076)
+- Support the new command `run`, used to run all the services generated
+
+### Changed
+
+- Fixed validator restart failure:
+  - Validator waits a grace period (2 epochs) before starting.
+  - Validator waits for the consensus sync to finish: verifying that the `/eth/v1/node/health` consensus endpoint returns with the HTTP code 200 in its docker-compose `healthcheck` condition so that the validator waits for the consensus be healthy. This replace the older track sync method.
+  - Validator service now always restarts unless stopped.
+  - Removed `v` from the tag on release scripts. Now we should use `Tag=v1.0.0` instead of `Tag=1.0.0`.
+
+### Removed
+
+- Removed config file dependency.
+- Removed `prysm` from consensus templates on Gnosis as supported client.
+
+## [v0.6.0] - 2022-12-23
+
+### Added
+
+- Support for Erigon as Execution Client.
+- Support for Besu as Execution Client.
+- Support for Windows.
+- Add search functionality to documentation using Algolia.
+
+### Changed
+
+- Completely replace keystore generation mechanism for Ethereum mainnet.
 
 ### Fixed
 
-- Fix inconsistent behavior of `AssignPorts` function test.
+- Revert Erigon Image from `v2.32.0` to `v2.29.0`.
+- Bad `{{end}}` tag on docker-compose_base template.
+- Inconsistent behavior of `AssignPorts` function test.
+- Dependencies install script bug for Ubuntu 22.04.
 
 ## [v0.5.1] - 2022-12-2